renamed LOFS header files to be preceeded by lofs- prefix. This is because some of the header files were overwriting system-wide and cuda header files.
