+++ include/lofs-dirstruct.h
