+++ include/lofs-macros.h
