+++ src/hdf2nc.c
@@ -1,7 +1,7 @@
 #include "../include/lofs-read.h"
-#include "../include/dirstruct.h"
-#include "../include/hdf2nc.h"
-#include "../include/limits.h"
+#include "../include/lofs-dirstruct.h"
+#include "../include/lofs-hdf2nc.h"
+#include "../include/lofs-limits.h"
 
 int main(int argc, char *argv[])
 {
